@@ -4,10 +4,11 @@ import kotlinx.coroutines.test.runTest
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatIOException
 import org.assertj.core.api.SoftAssertions
-import org.junit.jupiter.api.AfterAll
 import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.io.TempDir
 import ru.itis.kotlin.news_api.model.News
 import java.io.File
+import java.nio.file.Path
 import java.time.LocalDate
 import java.time.LocalTime
 import java.time.ZoneOffset
@@ -16,6 +17,9 @@ import kotlin.io.use
 private const val ZONE_OFFSET_ID = "+3"
 
 class KudaGoNewsApiTest {
+
+    @TempDir
+    private lateinit var tempDir:Path
     @Test
     fun `getNews should return list of first news`() {
         return runTest {
@@ -33,7 +37,56 @@ class KudaGoNewsApiTest {
     }
 
     @Test
-    fun `getMostRatedNews should return list of news that satisfy time period and sorted by rating`() {
+    fun `getMostRatedNews should return list of news that satisfy time period`() {
+        // given
+        val size = 3
+        val startDate = LocalDate.of(2024, 1, 15)
+        val endDate = LocalDate.of(2024, 9, 15)
+        var list = listOf(
+            News(1, "", null, "", "", 4, 0, 1710507332),
+            News(2, "", null, "", "", 2, 0, 1600000000),
+            News(3, "", null, "", "", 1, 0, 2000000000),
+            News(4, "", null, "", "", 3, 0, 1500000000),
+            News(5, "", null, "", "", 0, 0, 1700508832),
+            News(6, "", null, "", "", 5, 0, 1710507332),
+            News(7, "", null, "", "", 7, 0, 1600000000),
+            News(8, "", null, "", "", 3, 0, 1710534332),
+            News(9, "", null, "", "", 2, 0, 1500000000),
+            News(10, "", null, "", "", 1, 0, 1710508822)
+        )
+
+        // when
+        list = list.getMostRatedNews(size, startDate..endDate)
+
+        // then
+        val softAssertions = SoftAssertions()
+
+        softAssertions.assertThat(list.size)
+            .`as`("list.getMostRatedNews($size,startDate..endDate) method was called, but the list.size=${list.size}")
+            .isEqualTo(size)
+
+        for (news in list) {
+            softAssertions
+                .assertThat(news.publicationDate)
+                .`as` {
+                    "list.getMostRatedNews($size,$startDate..$endDate) method was called, " +
+                            "but the news in list have publicationDate=${news.publicationDate}"
+                }
+                .isBetween(
+                    startDate.toEpochSecond(
+                        LocalTime.MIN, ZoneOffset.of(
+                            ZONE_OFFSET_ID
+                        )
+                    ), endDate.toEpochSecond(LocalTime.MAX, ZoneOffset.of(ZONE_OFFSET_ID))
+                )
+
+        }
+
+        softAssertions.assertAll()
+    }
+
+    @Test
+    fun `getMostRatedNews should return list sorted by rating`() {
         // given
         val size = 3
         val startDate = LocalDate.of(2024, 1, 15)
@@ -63,19 +116,6 @@ class KudaGoNewsApiTest {
         var previousRating = Double.MAX_VALUE
         for (news in list) {
             softAssertions
-                .assertThat(news.publicationDate)
-                .`as` {
-                    "list.getMostRatedNews($size,$startDate..$endDate) method was called, " +
-                            "but the news in list have publicationDate=${news.publicationDate}"
-                }
-                .isBetween(
-                    startDate.toEpochSecond(
-                        LocalTime.MIN, ZoneOffset.of(
-                            ZONE_OFFSET_ID
-                        )
-                    ), endDate.toEpochSecond(LocalTime.MAX, ZoneOffset.of(ZONE_OFFSET_ID))
-                )
-            softAssertions
                 .assertThat(news.rating)
                 .`as`("The news is not in descending order in the list")
                 .isLessThanOrEqualTo(previousRating)
@@ -87,7 +127,48 @@ class KudaGoNewsApiTest {
     }
 
     @Test
-    fun `getMostRatedNewsAmongAllNews should return list of news that satisfy time period and sorted by rating`() {
+    fun `getMostRatedNewsAmongAllNews should return list of news that satisfy time period`() {
+        return runTest {
+            // given
+            val size = 50
+            val startDate = LocalDate.of(2024, 1, 15)
+            val endDate = LocalDate.of(2024, 9, 15)
+
+            // when
+            val list = getMostRatedNewsAmongAllNews(size, startDate..endDate)
+
+            // then
+            val softAssertions = SoftAssertions()
+
+            softAssertions
+                .assertThat(list.size)
+                .`as` {
+                    "list.getMostRatedNewsAmongAllNews($size,startDate..endDate) " +
+                            "method was called, but the list.size=${list.size}"
+                }
+                .isEqualTo(size)
+            for (news in list) {
+                softAssertions
+                    .assertThat(news.publicationDate)
+                    .`as` {
+                        "list.getMostRatedNewsAmongAllNews($size,$startDate..$endDate) method was called, " +
+                                "but the news in list have publicationDate=${news.publicationDate}"
+                    }
+                    .isBetween(
+                        startDate.toEpochSecond(
+                            LocalTime.MIN, ZoneOffset.of(
+                                ZONE_OFFSET_ID
+                            )
+                        ), endDate.toEpochSecond(LocalTime.MAX, ZoneOffset.of(ZONE_OFFSET_ID))
+                    )
+            }
+
+            softAssertions.assertAll()
+        }
+    }
+
+    @Test
+    fun `getMostRatedNewsAmongAllNews should return list of news sorted by rating`() {
         return runTest {
             // given
             val size = 50
@@ -110,19 +191,6 @@ class KudaGoNewsApiTest {
             var previousRating = Double.MAX_VALUE
             for (news in list) {
                 softAssertions
-                    .assertThat(news.publicationDate)
-                    .`as` {
-                        "list.getMostRatedNewsAmongAllNews($size,$startDate..$endDate) method was called, " +
-                                "but the news in list have publicationDate=${news.publicationDate}"
-                    }
-                    .isBetween(
-                        startDate.toEpochSecond(
-                            LocalTime.MIN, ZoneOffset.of(
-                                ZONE_OFFSET_ID
-                            )
-                        ), endDate.toEpochSecond(LocalTime.MAX, ZoneOffset.of(ZONE_OFFSET_ID))
-                    )
-                softAssertions
                     .assertThat(news.rating)
                     .`as`("The news is not in descending order in the list")
                     .isLessThanOrEqualTo(previousRating)
@@ -139,7 +207,7 @@ class KudaGoNewsApiTest {
         val list = listOf(
             News(1, "123", null, "description", "url", 1, 2, 90000000),
             News(2, "223", null, "description3", "url3", 5, 10, 1000000))
-        val path = "src/test/resources/saveNewsTest.csv"
+        val path = tempDir.resolve("saveNewsTest.csv").toString()
 
         // when
         saveNews(path, list)
@@ -166,7 +234,7 @@ class KudaGoNewsApiTest {
     @Test
     fun `saveNews should throw IOException if file already exist`() {
         // given
-        val path = "src/test/resources/alreadyExist.csv"
+        val path = tempDir.resolve("alreadyExist.csv").toString()
         File(path).createNewFile()
 
         // when
@@ -178,7 +246,7 @@ class KudaGoNewsApiTest {
     @Test
     fun `saveNews should throw IOException if file dont ends with csv`() {
         // given
-        val path = "file"
+        val path = tempDir.resolve("file").toString()
 
         // when
         assertThatIOException()
@@ -186,24 +254,4 @@ class KudaGoNewsApiTest {
             .isThrownBy { saveNews(path, listOf()) }
     }
 
-    companion object {
-        @JvmStatic
-        @AfterAll
-        fun closeClient() {
-            closeNewsClient()
-        }
-
-        @JvmStatic
-        @AfterAll
-        fun deleteTestFiles() {
-            val file1 = File("src/test/resources/saveNewsTest.csv")
-            if (file1.exists()) {
-                file1.delete()
-            }
-            val file2 = File("src/test/resources/alreadyExist.csv")
-            if (file2.exists()) {
-                file2.delete()
-            }
-        }
-    }
 }
